+++ src/pykdex/estimators/__init__.py
@@ -1,9 +1,9 @@
 # SPDX-FileCopyrightText: 2026 Jinghao Hu
 # SPDX-License-Identifier: MIT
 
-
 """User-facing density and intensity estimators."""
 
+from pykdex.estimators.network_kde import NetworkKDE
 from pykdex.estimators.spatial_kde import SpatialKDE
 
-__all__ = ["SpatialKDE"]
+__all__ = ["SpatialKDE", "NetworkKDE"]